chore(srw): extract http request function
