+++ packages/swr/src/index.ts
@@ -4,10 +4,6 @@ import {
   ClientDependenciesBuilder,
   ClientGeneratorsBuilder,
   ClientHeaderBuilder,
-  generateFormDataAndUrlEncodedFunction,
-  generateMutatorConfig,
-  generateMutatorRequestOptions,
-  generateOptions,
   generateVerbImports,
   GeneratorDependency,
   GeneratorMutator,
@@ -18,32 +14,14 @@ import {
   GetterProps,
   GetterPropType,
   GetterResponse,
-  isSyntheticDefaultImportsAllow,
   pascal,
   stringify,
   toObjectString,
   Verbs,
-  VERBS_WITH_BODY,
   jsDoc,
   SwrOptions,
 } from '@orval/core';
-
-const AXIOS_DEPENDENCIES: GeneratorDependency[] = [
-  {
-    exports: [
-      {
-        name: 'axios',
-        default: true,
-        values: true,
-        syntheticDefaultImport: true,
-      },
-      { name: 'AxiosRequestConfig' },
-      { name: 'AxiosResponse' },
-      { name: 'AxiosError' },
-    ],
-    dependency: 'axios',
-  },
-];
+import { AXIOS_DEPENDENCIES, generateSwrRequestFunction } from './client';
 
 const PARAMS_SERIALIZER_DEPENDENCIES: GeneratorDependency[] = [
   {
@@ -103,114 +81,6 @@ export const getSwrDependencies: ClientDependenciesBuilder = (
   ...SWR_INFINITE_DEPENDENCIES,
   ...SWR_MUTATION_DEPENDENCIES,
 ];
-
-const generateSwrRequestFunction = (
-  {
-    headers,
-    queryParams,
-    operationName,
-    response,
-    mutator,
-    body,
-    props,
-    verb,
-    formData,
-    formUrlEncoded,
-    override,
-    paramsSerializer,
-  }: GeneratorVerbOptions,
-  { route, context }: GeneratorOptions,
-) => {
-  const isRequestOptions = override?.requestOptions !== false;
-  const isFormData = override?.formData !== false;
-  const isFormUrlEncoded = override?.formUrlEncoded !== false;
-  const isExactOptionalPropertyTypes =
-    !!context.output.tsconfig?.compilerOptions?.exactOptionalPropertyTypes;
-  const isBodyVerb = VERBS_WITH_BODY.includes(verb);
-  const isSyntheticDefaultImportsAllowed = isSyntheticDefaultImportsAllow(
-    context.output.tsconfig,
-  );
-
-  const bodyForm = generateFormDataAndUrlEncodedFunction({
-    formData,
-    formUrlEncoded,
-    body,
-    isFormData,
-    isFormUrlEncoded,
-  });
-
-  if (mutator) {
-    const mutatorConfig = generateMutatorConfig({
-      route,
-      body,
-      headers,
-      queryParams,
-      response,
-      verb,
-      isFormData,
-      isFormUrlEncoded,
-      hasSignal: false,
-      isBodyVerb,
-      isExactOptionalPropertyTypes,
-    });
-
-    const propsImplementation =
-      mutator?.bodyTypeName && body.definition
-        ? toObjectString(props, 'implementation').replace(
-            new RegExp(`(\\w*):\\s?${body.definition}`),
-            `$1: ${mutator.bodyTypeName}<${body.definition}>`,
-          )
-        : toObjectString(props, 'implementation');
-
-    const requestOptions = isRequestOptions
-      ? generateMutatorRequestOptions(
-          override?.requestOptions,
-          mutator.hasSecondArg,
-        )
-      : '';
-
-    return `export const ${operationName} = (\n    ${propsImplementation}\n ${
-      isRequestOptions && mutator.hasSecondArg
-        ? `options${context.output.optionsParamRequired ? '' : '?'}: SecondParameter<typeof ${mutator.name}>`
-        : ''
-    }) => {${bodyForm}
-      return ${mutator.name}<${response.definition.success || 'unknown'}>(
-      ${mutatorConfig},
-      ${requestOptions});
-    }
-  `;
-  }
-
-  const options = generateOptions({
-    route,
-    body,
-    headers,
-    queryParams,
-    response,
-    verb,
-    requestOptions: override?.requestOptions,
-    isFormData,
-    isFormUrlEncoded,
-    paramsSerializer,
-    paramsSerializerOptions: override?.paramsSerializerOptions,
-    isExactOptionalPropertyTypes,
-    hasSignal: false,
-  });
-
-  return `export const ${operationName} = (\n    ${toObjectString(
-    props,
-    'implementation',
-  )} ${
-    isRequestOptions ? `options?: AxiosRequestConfig\n` : ''
-  } ): Promise<AxiosResponse<${
-    response.definition.success || 'unknown'
-  }>> => {${bodyForm}
-    return axios${
-      !isSyntheticDefaultImportsAllowed ? '.default' : ''
-    }.${verb}(${options});
-  }
-`;
-};
 
 const generateSwrArguments = ({
   operationName,
